Add ability to create contanct sucessfully
