+++ packages/esm-patient-clinical-view-app/src/config-schema.ts
@@ -96,7 +96,7 @@ export const configSchema = {
   openmrsIdentifierSourceUuid: {
     _type: Type.String,
     _description: 'OpenMRS Identifier Source UUID (Identifier Generator for OpenMRS ID)',
-    _default: 'fb034aac-2353-4940-abe2-7bc94e7c1e71',
+    _default: '1952cc86-4f48-4737-a0ef-5e8a5bb63e41',
   },
   hivProgramUuid: {
     _type: Type.String,
