+++ packages/esm-patient-clinical-view-app/src/contact-list/contact-list.resource.tsx
@@ -3,91 +3,8 @@ import { useMemo } from 'react';
 import useSWR from 'swr';
 import { z } from 'zod';
 import { ConfigObject } from '../config-schema';
+import { Enrollment, HTSEncounter, Contact, Patient, Person, Relationship, RelationShipType } from '../types';
 
-export interface Relationship {
-  display: string;
-  uuid: string;
-  personA: Person;
-  personB: Person;
-  relationshipType: {
-    uuid: string;
-    display: string;
-    aIsToB: string;
-    bIsToA: string;
-  };
-  startDate: string;
-}
-
-interface Contact {
-  uuid: string;
-  name: string;
-  display: string;
-  relativeAge: number;
-  dead: boolean;
-  causeOfDeath: string;
-  relativeUuid: string;
-  relationshipType: string;
-  patientUuid: string;
-  gender: string;
-  contact: string | null;
-  startDate: string | null;
-  baselineHIVStatus: string | null;
-  personContactCreated: string | null;
-  livingWithClient: string | null;
-  pnsAproach: string | null;
-}
-
-interface Person {
-  uuid: string;
-  age: number;
-  dead: boolean;
-  display: string;
-  causeOfDeath: string;
-  gender: string;
-  attributes: {
-    uuid: string;
-    display: string;
-    attributeType: {
-      uuid: string;
-      display: string;
-    };
-  }[];
-}
-
-interface Patient {
-  uuid: string;
-  person: Person;
-  identifiers: {
-    uuid: string;
-  }[];
-}
-
-interface RelationShipType {
-  uuid: string;
-  displayAIsToB: string;
-}
-
-interface Enrollment {
-  uuid: string;
-  program: {
-    name: string;
-    uuid: string;
-  };
-}
-
-interface HTSEncounter {
-  uuid: string;
-  display: string;
-  encounterDatetime: string;
-  obs: {
-    uuid: string;
-    display: string;
-    value: {
-      uuid: string;
-      display: string;
-    };
-  }[];
-}
 export const ContactListFormSchema = z.object({
   listingDate: z.date({ coerce: true }),
   givenName: z.string().min(1, 'Required'),
