+++ packages/esm-patient-clinical-view-app/src/types/index.ts
@@ -62,3 +62,87 @@ export interface ConceptToFormLabelMap {
   display: string;
   answers: null | Array<string>;
 }
+export interface Relationship {
+  display: string;
+  uuid: string;
+  personA: Person;
+  personB: Person;
+  relationshipType: {
+    uuid: string;
+    display: string;
+    aIsToB: string;
+    bIsToA: string;
+  };
+  startDate: string;
+}
+
+export interface Contact {
+  uuid: string;
+  name: string;
+  display: string;
+  relativeAge: number;
+  dead: boolean;
+  causeOfDeath: string;
+  relativeUuid: string;
+  relationshipType: string;
+  patientUuid: string;
+  gender: string;
+  contact: string | null;
+  startDate: string | null;
+  baselineHIVStatus: string | null;
+  personContactCreated: string | null;
+  livingWithClient: string | null;
+  pnsAproach: string | null;
+}
+
+export interface Person {
+  uuid: string;
+  age: number;
+  dead: boolean;
+  display: string;
+  causeOfDeath: string;
+  gender: string;
+  attributes: {
+    uuid: string;
+    display: string;
+    attributeType: {
+      uuid: string;
+      display: string;
+    };
+  }[];
+}
+
+export interface Patient {
+  uuid: string;
+  person: Person;
+  identifiers: {
+    uuid: string;
+  }[];
+}
+
+export interface RelationShipType {
+  uuid: string;
+  displayAIsToB: string;
+}
+
+export interface Enrollment {
+  uuid: string;
+  program: {
+    name: string;
+    uuid: string;
+  };
+}
+
+export interface HTSEncounter {
+  uuid: string;
+  display: string;
+  encounterDatetime: string;
+  obs: {
+    uuid: string;
+    display: string;
+    value: {
+      uuid: string;
+      display: string;
+    };
+  }[];
+}
